fix error permission_error views

+++ users/views.py
@@ -7,6 +7,9 @@ from .forms import *
 
 def view_coordenadoresCurso(request):
     try:
+        if not request.user.is_authenticated:
+            return render(request, 'permission_error.html')
+
         if request.user.tipo_usuario == 'CoordenadorEnsino':
             data = {}
             form = UserCreationForm(request.POST or None)
@@ -34,6 +37,9 @@ def view_coordenadoresCurso(request):
 
 def ver_coordenadorCurso(request, id_coordenadorcurso):
     try:
+        if not request.user.is_authenticated:
+            return render(request, 'permission_error.html')
+
         if request.user.tipo_usuario == 'CoordenadorEnsino':
             data = {}
             coordenadorCurso = User.objects.get(id=id_coordenadorcurso)
@@ -53,6 +59,9 @@ def ver_coordenadorCurso(request, id_coordenadorcurso):
 
 def delete_coordenadorCurso(request, id_coordenadorcurso):
     try:
+        if not request.user.is_authenticated:
+            return render(request, 'permission_error.html')
+
         if request.user.tipo_usuario == 'CoordenadorEnsino':
             data = {'mensagem': "Coordenador de Curso " + str(id_coordenadorcurso) + " removido com sucesso!"}
             coordenadorCurso = User.objects.get(id=id_coordenadorcurso)
@@ -64,31 +73,30 @@ def delete_coordenadorCurso(request, id_coordenadorcurso):
     except:
         data = {'mensagem': "Não foi possível excluir o Coordenador de Curso!"}
         return render(request, 'users/coordenadorEnsino/error.html', data)
-
-
-
 
 
 def delete_coordenadorCurso(request, id_coordenadorcurso):
     try:
+        if not request.user.is_authenticated:
+            return render(request, 'permission_error.html')
+
         if request.user.tipo_usuario == 'CoordenadorEnsino':
             data = {'mensagem': "Coordenador de Curso " + str(id_coordenadorcurso) + " removido com sucesso!"}
             coordenadorCurso = User.objects.get(id=id_coordenadorcurso)
-            n = coordenadorCurso.delete()
-            print("Removidos: ", n)
+            coordenadorCurso.delete()
             return render(request, 'users/coordenadorEnsino/cadastro_sucesso.html', data)
         else:
-            return render(request, '../../templates/permission_error.html')
+            return render(request, 'permission_error.html')
     except:
         data = {'mensagem': "Não foi possível excluir o Coordenador de Curso!"}
         return render(request, 'users/coordenadorEnsino/error.html', data)
 
 
-
-
-
 def view_professores(request):
     try:
+        if not request.user.is_authenticated:
+            return render(request, 'permission_error.html')
+
         if request.user.tipo_usuario == 'CoordenadorCurso':
             data = {}
             form = UserCreationForm(request.POST or None)
@@ -108,7 +116,7 @@ def view_professores(request):
             data['form'] = form
             return render(request, 'users/coordenadorCurso/professores.html', data)
         else:
-            return render(request, '../../templates/permission_error.html')
+            return render(request, '.permission_error.html')
     except:
         data = {'mensagem': "Ocorreu um erro interno!" }
         return render(request, 'users/coordenadorCurso/error.html', data)
@@ -116,6 +124,9 @@ def view_professores(request):
 
 def ver_professor(request, id_professor):
     try:
+        if not request.user.is_authenticated:
+            return render(request, 'permission_error.html')
+
         if request.user.tipo_usuario == 'CoordenadorCurso':
             data = {}
             professor = User.objects.get(id=id_professor)
@@ -126,7 +137,7 @@ def ver_professor(request, id_professor):
                 data = {'mensagem': "Não foi possível localizar o Professor!"}
                 return render(request, 'users/coordenadorCurso/error.html', data)
         else:
-            return render(request, '../../templates/permission_error.html')
+            return render(request, 'permission_error.html')
     except:
         data = {'mensagem': "Não foi possível visualizar o Professor!"}
         return render(request, 'users/coordenadorCurso/error.html', data)
@@ -134,14 +145,16 @@ def ver_professor(request, id_professor):
 
 def delete_professor(request, id_professor):
     try:
+        if not request.user.is_authenticated:
+            return render(request, 'permission_error.html')
+
         if request.user.tipo_usuario == 'CoordenadorCurso':
             data = {'mensagem': "Professor " + str(id_professor) + " removido com sucesso!"}
             professor = User.objects.get(id=id_professor)
-            n = professor.delete()
-            print("Removidos: ", n)
+            professor.delete()
             return render(request, 'users/coordenadorCurso/cadastro_sucesso.html', data)
         else:
-            return render(request, '../../templates/permission_error.html')
+            return render(request, 'permission_error.html')
     except:
         data = {'mensagem': "Não foi possível excluir o Professor!"}
         return render(request, 'users/coordenadorCurso/error.html', data)
